# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Monitor",
        "contact": {
            "name": "Glassen_IT"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://62.113.103.54/api",
            "description": "API Monitor"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Регестрация нового пользователя",
                "operationId": "5d4fa5d5607ed675ba93907e1fc94534",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "login": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "role": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "login": "user1",
                                    "password": "password1",
                                    "email": "example@gmail.com",
                                    "role": "0|1|2"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "login": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        },
                                        "role": {
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string"
                                        },
                                        "id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "login": "testmanager",
                                        "email": "testexample@mail.ru",
                                        "role": "2",
                                        "updated_at": "2023-06-26T11:06:30.000000Z",
                                        "created_at": "2023-06-26T11:06:30.000000Z",
                                        "id": 4
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResult401"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Неверные входные значения",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResult422"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Аутентификация пользователя",
                "operationId": "67d6d0fdd54fba132e1b6f55ca254183",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "login": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "login": "user1",
                                    "password": "password1"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "login": {
                                                        "type": "string"
                                                    },
                                                    "email": {
                                                        "type": "string"
                                                    },
                                                    "role": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "login": "admin",
                                                    "email": "anton_starilov@mail.ru",
                                                    "role": "1"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Неверный логин или пароль",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResult422"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResult422": {
                "title": "Шаблон ошибки",
                "properties": {
                    "code": {
                        "type": "integer"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "code": 422,
                    "message": "Неверные входные значения"
                }
            },
            "ErrorResult401": {
                "title": "Шаблон ошибки статуса авторизации",
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "message": "Unauthenticated."
                }
            }
        }
    },
...
}